requirements.txt
# packages in environment at /esat/pyrite/svandenh/hal/miniconda3/envs/pytorch:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
apex                      0.1                      pypi_0    pypi
blas                      1.0                    openblas  
brotlipy                  0.7.0           py38h140841e_1003  
bzip2                     1.0.8                h7b6447c_0  
ca-certificates           2020.12.8            h6ffa863_0  
certifi                   2020.12.5        py38h6ffa863_0  
cffi                      1.14.2           py38he30daa8_0  
chardet                   4.0.0           py38h6ffa863_1003  
click                     7.1.2                      py_0  
cloudpickle               1.6.0                      py_0  
cmake                     3.14.0               h52cb24c_0  
cryptography              3.3.1            py38hb567c45_0  
cycler                    0.10.0                   py38_0  
cython                    0.29.21                  pypi_0    pypi
cytoolz                   0.11.0           py38h7b6447c_0  
dask-core                 2.30.0                     py_0  
decorator                 4.4.2                      py_0  
easydict                  1.9                        py_0    conda-forge
einops                    0.3.0                      py_0    conda-forge
expat                     2.2.9                he6710b0_2  
freetype                  2.10.2               h5ab3b9f_0  
h5py                      2.10.0           py38h7918eee_0  
hdf5                      1.10.4               hb1b8bf9_0  
idna                      2.10                       py_0  
imageio                   2.9.0                      py_0  
joblib                    0.16.0                     py_0  
jpeg                      9b                   hcb7ba68_2  
kiwisolver                1.2.0            py38hfd86e86_0  
krb5                      1.18.2               h597af5e_0  
lcms2                     2.11                 h396b838_0  
ld_impl_linux-ppc64le     2.33.1               h0f24833_7  
libcurl                   7.71.1               h20c2e04_1  
libedit                   3.1.20191231         h14c3975_1  
libffi                    3.3                  he6710b0_2  
libgcc-ng                 8.2.0                h822a55f_1  
libgfortran-ng            7.3.0                h822a55f_1  
libopenblas               0.3.10               h5a2b251_0  
libpng                    1.6.37               hbc83047_0  
libssh2                   1.9.0                h1ba5d50_1  
libstdcxx-ng              8.2.0                h822a55f_1  
libtiff                   4.1.0                h2733197_1  
lz4-c                     1.9.2                he6710b0_1  
matplotlib                3.3.1                         0  
matplotlib-base           3.3.1            py38h8254106_0  
ncurses                   6.2                  he6710b0_1  
networkx                  2.5                        py_0  
numpy                     1.19.1           py38h30dfecb_0  
numpy-base                1.19.1           py38h75fe3a5_0  
olefile                   0.46                       py_0  
openssl                   1.1.1i               h140841e_0  
pillow                    7.2.0            py38haac5956_0  
pip                       20.2.2                   py38_0  
pycocotools               2.0.2                    pypi_0    pypi
pycparser                 2.20                       py_2  
pyopenssl                 20.0.1             pyhd3eb1b0_1  
pyparsing                 2.4.7                      py_0  
pyquaternion              0.9.9                    pypi_0    pypi
pysocks                   1.7.1            py38h6ffa863_0  
python                    3.8.5                h3a5dbe6_1  
python-dateutil           2.8.1                      py_0  
pywavelets                1.1.1            py38h7b6447c_2  
pyyaml                    5.3.1            py38h7b6447c_1  
readline                  8.0                  h7b6447c_0  
requests                  2.25.1             pyhd3eb1b0_0  
rhash                     1.3.8                h1ba5d50_0  
scikit-image              0.17.2           py38hdf5156a_0  
scikit-learn              0.23.2           py38h0573a6f_0  
scipy                     1.5.2            py38habc2bb6_0  
setuptools                49.6.0                   py38_0  
six                       1.15.0                   pypi_0    pypi
sqlite                    3.33.0               hbc83047_0  
termcolor                 1.1.0                    py38_1  
threadpoolctl             2.1.0              pyh5ca1d4c_0  
tifffile                  2020.10.1        py38hdd07704_2  
tk                        8.6.10               hbc83047_0  
toolz                     0.11.1                     py_0  
torch                     1.4.0a0+7f73f1d          pypi_0    pypi
torchvision               0.5.0a0+85b8fbf          pypi_0    pypi
tornado                   6.0.4            py38h7b6447c_1  
tqdm                      4.50.2                     py_0  
typing                    3.7.4.3                  py38_0  
urllib3                   1.26.2             pyhd3eb1b0_0  
wheel                     0.35.1                     py_0  
xz                        5.2.5                h7b6447c_0  
yaml                      0.2.5                h7b6447c_0  
zlib                      1.2.11               h7b6447c_3  
zstd                      1.4.5                h9ceee32_0  
